ML/requirements.txt
opencv-python>=4.5.0
mediapipe>=0.8.9
tensorflow>=2.8.0
numpy>=1.21.0
Pillow>=8.3.0
scikit-image>=0.18.0
matplotlib>=3.5.0
flask>=2.0.0
flask-cors>=3.0.10
requests>=2.26.0